ex7.8: Fix ordering for 3 or more comparisons
It seems I got confused and was reversing the slice of comparisons
instead of just prepending the new one into the most significant
position. Thanks to covrom (Roman TSovanyan) for pointing this out.

Also, push the decision for how many max comparisons to allow up to the
client, since there could be more columns available to sort on than
fields in a struct.