Improve example 8.1
This commit improves both the clock server and the wallclock client
from example 8.1.

clock.go is updated to add a feature that reads from an environment
variable 'TZ' to set the time zone of the time server from one
of the defined IANA time database timezone definitions.  This allows
each clock to report in its assigned time zone without configuration
by the client, and is in keeping with the instructions for EX 8.1
which suggests the use of a 'TZ' variable when launching the
clock

clockwall.go is rewritten to implment the in-place updates typically
seen on a wall of clocks.  The original provided concurrent updates
from several clock servers, but the input scrolled down the screen
and did not reflect the use case of a clock wall that visually
tracks times in multiple time zones as typically seen in some hotels,
command and control centers, and data centers with worldwide
customers.  The updated clockwall.go provides continuous feeds from
clock servers on a single line in keeping with the idea of a clock
wall.